+++ crates/formatter/src/formatter.rs
@@ -435,7 +435,7 @@ impl VerylWalker for Formatter {
 
     /// Semantic action for non-terminal 'Type'
     fn r#type(&mut self, arg: &Type) {
-        if let Some(ref x) = arg.type_opt {
+        for x in &arg.type_list {
             self.type_modifier(&x.type_modifier);
             self.space(1);
         }
@@ -444,7 +444,7 @@ impl VerylWalker for Formatter {
             TypeGroup::ScopedIdentifier(x) => self.scoped_identifier(&x.scoped_identifier),
         };
         self.space(1);
-        for x in &arg.type_list {
+        for x in &arg.type_list0 {
             self.width(&x.width);
         }
     }