+++ crates/parser/src/veryl_token.rs
@@ -238,6 +238,7 @@ token_with_comments!(PosedgeToken, PosedgeTerm, posedge_term);
 token_with_comments!(RefToken, RefTerm, ref_term);
 token_with_comments!(RepeatToken, RepeatTerm, repeat_term);
 token_with_comments!(ReturnToken, ReturnTerm, return_term);
+token_with_comments!(SignedToken, SignedTerm, signed_term);
 token_with_comments!(StepToken, StepTerm, step_term);
 token_with_comments!(StructToken, StructTerm, struct_term);
 token_with_comments!(SyncHighToken, SyncHighTerm, sync_high_term);